2003-08-24  Martyn Russell  <ginxd@btopenworld.com>
	* loudmouth/lm-connection.c:
	(connection_in_event): Use an int for bytes_read. Since bytes_read is
	  now also used for gnutls return value (which can be < 0) a gsize
	  (unsigned) is not a good idea...

+++ loudmouth/lm-connection.c
@@ -374,7 +374,7 @@ connection_in_event (GIOChannel   *source,
 		     LmConnection *connection)
 {
 	gchar     buf[IN_BUFFER_SIZE];
-	gsize     bytes_read;
+	gint      bytes_read;
 	GIOStatus status;
        
 	if (!connection->io_channel) {
@@ -407,7 +407,6 @@ connection_in_event (GIOChannel   *source,
 
 	if (status != G_IO_STATUS_NORMAL) {
 		gint reason;
-		
 		
 		switch (status) {
 		case G_IO_STATUS_EOF:
